README.md: propose `U-Boot-Dock`
Intend: Proposal

Motivation:
The original project name is a bit war movie centric.
I'd like to propose that a U-Boot's original interpretation is more
`boot-loader` oriented.
But the nautic correlation is inteded and well suited. Why not using
this metapher and give U-Boot a dock:
create and revamp its productive configuration and environment?

Signed-off-by: Ralf Zerres <ralf.zerres@networkx.de>